2006-08-15  Steve Falkenburg  <sfalken@apple.com>
        Reviewed by mjs.

        Build fix.

        * JavaScriptCore.vcproj/JavaScriptCore/JavaScriptCore.vcproj:
        * wtf/Assertions.h:



Canonical link: https://commits.webkit.org/13326@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@15906 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/wtf/Assertions.h
@@ -96,7 +96,7 @@ void WTFLog(const char *file, int line, const char *function, WTFLogChannel *cha
 
 // CRASH -- gets us into the debugger or the crash reporter -- signals are ignored by the crash reporter so we must do better
 
-#define CRASH() *(int *)0xbbadbeef = 0
+#define CRASH() *(int *)(uintptr_t)0xbbadbeef = 0
 
 // ASSERT, ASSERT_WITH_MESSAGE, ASSERT_NOT_REACHED
